// services/account-service/src/proto/account.pb.ts
// Generated by ts-proto v1.0.0
// DO NOT EDIT

import * as grpc from '@grpc/grpc-js';

export interface GetUserRequest {
  id: number;
}

export interface CreateUserRequest {
  mobile: string;
  nickname: string;
  avatarUrl: string;
}

export interface UpdateUserRequest {
  id: number;
  nickname: string;
  avatarUrl: string;
}

export interface CreateAddressRequest {
  userId: number;
  consignee: string;
  mobile: string;
  province: string;
  city: string;
  district: string;
  detail: string;
  isDefault: boolean;
}

export interface GetAddressesRequest {
  userId: number;
}

export interface UpdateAddressRequest {
  id: number;
  consignee: string;
  mobile: string;
  province: string;
  city: string;
  district: string;
  detail: string;
  isDefault: boolean;
}

export interface DeleteAddressRequest {
  id: number;
}

export interface UserResponse {
  id: number;
  mobile: string;
  nickname: string;
  avatarUrl: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface AddressResponse {
  id: number;
  userId: number;
  consignee: string;
  mobile: string;
  province: string;
  city: string;
  district: string;
  detail: string;
  isDefault: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AddressListResponse {
  addresses: AddressResponse[];
}

export interface Empty {}

export interface AccountServiceHandlers extends grpc.UntypedServiceImplementation {
  getUser: grpc.handleUnaryCall<GetUserRequest, UserResponse>;
  createUser: grpc.handleUnaryCall<CreateUserRequest, UserResponse>;
  updateUser: grpc.handleUnaryCall<UpdateUserRequest, UserResponse>;
  createAddress: grpc.handleUnaryCall<CreateAddressRequest, AddressResponse>;
  getAddresses: grpc.handleUnaryCall<GetAddressesRequest, AddressListResponse>;
  updateAddress: grpc.handleUnaryCall<UpdateAddressRequest, AddressResponse>;
  deleteAddress: grpc.handleUnaryCall<DeleteAddressRequest, Empty>;
}

export const AccountServiceService: grpc.ServiceDefinition<AccountServiceHandlers> = {} as any;